# docs/reference/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rohy API",
    "version": "1.0.0",
    "description": "Auto-generated from server/routes/*.js by scripts/docs-gen/gen-api.mjs. Do not hand-edit. Regenerate with `npm run docs:gen:api`."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT in Authorization: Bearer <token>."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "HttpOnly JWT cookie. State-changing cookie-authenticated requests also require the X-CSRF-Token header."
      }
    }
  },
  "paths": {
...
    "/api/plugins/{pluginId}/*splat": {
      "get": {
        "summary": "Get /plugins/:pluginId/*splat",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error (envelope via server/redaction.js)"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "x-rohy-source": "server/routes/plugins-routes.js:346",
        "x-rohy-auth": [
          "authenticateToken",
          "requireStudent"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/plugins/{pluginId}/catalog": {
      "get": {
        "summary": "Get /plugins/:pluginId/catalog",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error (envelope via server/redaction.js)"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "x-rohy-source": "server/routes/plugins-routes.js:130",
        "x-rohy-auth": [
          "authenticateToken"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/plugins/{pluginId}/settings": {
      "get": {
        "summary": "Get /plugins/:pluginId/settings",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error (envelope via server/redaction.js)"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "x-rohy-source": "server/routes/plugins-routes.js:259",
        "x-rohy-auth": [
          "authenticateToken"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "summary": "Replace /plugins/:pluginId/settings",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation error (envelope via server/redaction.js)"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient role"
          }
        },
        "x-rohy-source": "server/routes/plugins-routes.js:288",
        "x-rohy-auth": [
          "authenticateToken"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  }
}